// src/utils/interface.ts
import { Request } from "express";
import { Types } from "mongoose";
import { User } from "../models/user.model";

export interface IUserToken
  extends Omit<
    User,
    | "password"
    | "activationCode"
    | "isActive"
    | "email"
    | "fullName"
    | "profilePicture"
    | "username"
  > {
  id?: Types.ObjectId;
}

export interface IReqUser extends Request {
  user?: IUserToken;
}

export interface iPaginationQuery {
  page: number;
  limit: number;
  search?: string;
}